+++ src/components/index.ts
@@ -1,4 +1,3 @@
-export * from "./Layer";
 export * from "./Button";
 export * from "./Accordion";
 export * from "./Avatar";