+++ CHANGELOG.md
@@ -4,7 +4,7 @@
 
 ### Minor Changes
 
-- [#346](https://github.com/borderux/recursica/pull/346) [`ea86688`](https://github.com/borderux/recursica/commit/ea866884a4ed1eb84eed5741bf0329940a683619) Thanks [@mlmassey](https://github.com/mlmassey)! - Updated for latest CSS config and postcss processing
+- ea86688: Updated for latest CSS config and postcss processing
 
 ## 0.2.0